Populate sync env groups inside key value array.
